- Patch #19697 by Morbus: FOLKSONOMY.
This patch adds folksonomy support to Drupal (named internally as "Free tagging"). In a nutshell, the core difference is the input method: unlike normal taxonomies which are administratively controlled, a "free tagging" vocabulary allows tag creation when the node is submitted. It does this through an text input box, as opposed to a dropdown or selectbox. This patch:

    * Removes the useless "Preview form" of a vocabulary.
    * Alters the vocabulary table to include a new "tags" column.
    * Adds a new "Free tagging" preference on vocabulary creation/editing.
    * Modifies the vocabulary overview to support pagers for free tagging vocabs.

The new code integrates tightly with the existing taxonomy code. The only additional processing occurs on node save and edit, where we parse through the tags associated with a node. All other display (and thus, code) remains the same.